+++ cli/internal/helm/generateCilium.sh
@@ -5,9 +5,16 @@ shopt -s inherit_errexit
 
 calldir=$(pwd)
 ciliumTmpDir=$(mktemp -d)
-cd "${ciliumTmpDir}" || exit 1
-git clone --depth 1 -b 1.12.1 https://github.com/cilium/cilium.git
-cd cilium || exit 1
+pushd "${ciliumTmpDir}"
+git clone --filter=blob:none --no-checkout --sparse --depth 1 -b 1.12.1 https://github.com/cilium/cilium.git
+pushd cilium
+
+git sparse-checkout add install/kubernetes/cilium
+git checkout
+
 git apply "${calldir}"/cilium.patch
 cp -r install/kubernetes/cilium "${calldir}"/charts
+
+popd
+popd
 rm -r "${ciliumTmpDir}"